refactor(auth): isolate owner-aware identity probes
